# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SwarmUI API
  description: |
    API documentation for SwarmUI - AI-powered development environment with parallel CLI agent orchestration.
    
    ## Overview
    SwarmUI provides a comprehensive API for managing:
    - Chat sessions with AI agents
    - Project and ticket management
    - Background job processing
    - Git operations
    - File system operations
    - Terminal sessions
    - Testing and debugging
    
    ## Authentication
    Most endpoints require authentication via NextAuth.js session cookies.
    
    ## WebSocket
    Real-time updates are available via WebSocket at `ws://localhost:3000`.
    See [WebSocket API documentation](./WEBSOCKET_API.md) for details.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: SwarmUI Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:4000
    description: Alternative dev port
...
paths:
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List all projects
      operationId: listProjects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get project by ID
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    patch:
      tags: [Projects]
      summary: Update project
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Project updated
    delete:
      tags: [Projects]
      summary: Delete project
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
...
  /api/projects/{id}/epics:
    get:
      tags: [Epics]
      summary: List epics for a project
      operationId: listProjectEpics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of epics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Epic'
    post:
      tags: [Epics]
      summary: Create an epic
      operationId: createEpic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EpicCreate'
      responses:
        '201':
          description: Epic created
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
  schemas:
...
    AgentRole:
      type: string
      enum: [researcher, planner, coder, validator, security, synthesizer]
    CLIProvider:
      type: string
      enum: [cursor, gemini, claude, copilot, codex, rovo, custom]
...
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        prd:
          type: string
        prdStatus:
          type: string
          enum: [draft, approved, rejected]
        features:
          type: array
          items:
            type: string
        epics:
          type: array
          items:
            $ref: '#/components/schemas/Epic'
        tickets:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
        createdAt:
          type: integer
        updatedAt:
          type: integer
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, in_progress, completed, archived]
    Epic:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        ticketIds:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, active, completed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        createdAt:
          type: integer
        updatedAt:
          type: integer
    EpicCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
    Ticket:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        title:
          type: string
        description:
          type: string
        acceptanceCriteria:
          type: array
          items:
            type: string
        complexity:
          $ref: '#/components/schemas/TicketComplexity'
        status:
          $ref: '#/components/schemas/TicketStatus'
        level:
          $ref: '#/components/schemas/TicketLevel'
        assignedRole:
          $ref: '#/components/schemas/AgentRole'
        assignedProvider:
          $ref: '#/components/schemas/CLIProvider'
        parentId:
          type: string
        epicId:
          type: string
        dependencies:
          type: array
          items:
            type: string
        output:
          type: string
        diff:
          type: string
        confidence:
          type: number
        sla:
          $ref: '#/components/schemas/SLAConfig'
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
    TicketComplexity:
      type: string
      enum: [S, M, L, XL]
    TicketStatus:
      type: string
      enum: [backlog, in_progress, review, approved, rejected, done]
    TicketLevel:
      type: string
      enum: [feature, epic, story, task, subtask, subatomic]
    SLAConfig:
      type: object
      properties:
        responseTimeHours:
          type: number
        resolutionTimeHours:
          type: number
        priority:
          type: string
          enum: [low, medium, high, critical]
...